requirements.txt
feedparser
google-generativeai
python-dotenv
beautifulsoup4